@@ -7,6 +7,7 @@
     "@jest-mock/express": "^1.4.5",
     "@types/cors": "^2.8.11",
     "@types/express": "^4.17.13",
+    "@types/express-serve-static-core": "4.17.30",
     "@types/faker": "^6.6.9",
     "@types/jest": "^27.0.2",
     "@types/jsonwebtoken": "^8.5.5",
@@ -26,12 +27,13 @@
     "prettier": "^2.7.1",
     "supertest": "^6.1.6",
     "ts-jest": "^27.0.5",
-    "ts-node": "^10.1.0",
-    "typescript": "3.3.3333"
+    "ts-node": "^10.9.1",
+    "typescript": "^4.7.4"
   },
   "dependencies": {
     "@sendgrid/mail": "^7.4.7",
     "@types/node-cron": "^3.0.1",
+    "@types/node-schedule": "",
     "archiver": "^5.3.0",
     "body-parser": "^1.19.0",
     "class-validator": "^0.13.1",
@@ -48,12 +50,14 @@
     "keccak": "^3.0.1",
     "moment": "^2.29.1",
     "multiformats": "^9.4.10",
-    "node-cron": "^3.0.0",
+    "node-cron": "^3.0.2",
+    "node-schedule": "",
     "pg": "^8.4.0",
     "pino": "^7.0.5",
+    "pm2": "^5.2.2",
     "react-select-country-list": "^2.2.3",
     "reflect-metadata": "^0.1.10",
-    "typeorm": "0.2.34",
+    "typeorm": "^0.2.34",
     "typeorm-seeding": "^1.6.1",
     "uint8arrays": "^3.0.0",
     "uuid": "^8.3.2",
@@ -61,16 +65,16 @@
   },
   "scripts": {
     "start": "export NODE_ENV=development&&nodemon",
-    "build": "export NODE_ENV=production&&nodemon",
+    "build": "tsc -p tsconfig.json --noEmit false",
     "test": "jest",
-    "typeorm": "node --require ts-node/register ./node_modules/typeorm/cli.js",
+    "typeorm": "node --experimental-modules --es-module-specifier-resolution=node --loader ts-node/esm -r ts-node/register ./node_modules/typeorm/cli.js",
     "format": "prettier --write 'src/**/*.{ts}'",
     "seed:config": "ts-node ./node_modules/typeorm-seeding/dist/cli.js config",
     "seed:run": "ts-node ./node_modules/typeorm-seeding/dist/cli.js seed",
-    "schema:drop": "ts-node ./node_modules/typeorm/cli.js schema:drop",
-    "sc:dr": "ts-node ./node_modules/typeorm/cli.js schema:drop",
-    "schema:sync": "ts-node ./node_modules/typeorm/cli.js schema:sync",
-    "sc:sy": "ts-node ./node_modules/typeorm/cli.js schema:sync"
+    "schema:drop": "yarn typeorm schema:drop",
+    "sc:dr": "yarn schema:drop",
+    "schema:sync": "yarn typeorm schema:sync",
+    "sc:sy": "yarn schema:sync"
   },
   "lint-staged": {
     "src/**/*.{ts}": [
@@ -83,5 +87,15 @@
       "pre-commit": "lint-staged"
     }
   },
-  "private": true
+  "exclude": [
+    "test/**/*",
+    "node_modules",
+    "src/database/*"
+  ],
+  "exports": {
+    "require": "./index.cjs",
+    "import": "./index.js"
+  },
+  "private": true,
+  "type": "module"
 }
