@@ -1,6 +1,6 @@
 import { Request, Response } from 'express';
 import * as IPFS from 'ipfs-core';
-import all = require('it-all');
+import all from 'it-all'
 
 export const get_ipfs_file = async (request: Request, response: Response) => {
   const {
